+++ NEWS.md
@@ -1,8 +1,14 @@
 # JointAI (development version)
 
+## Bug fixes
+* Error when using the same function twice in a model formula with different 
+  number of variables (e.g., I(x^2) and I(a/b)) fixed.
+
+## Small improvements
 * `all_vars()`: update of the function; can now handle an unspecified number of
   input objects and can extract variable names from formulas and character 
   strings that are valid variable names
+
 
 --------------------------------------------------------------------------------